122414 D3X9 MovingCube and Grid casting Float

+++ DirectX9/DX3D9HomeAssignment/122213 DrawingSimpleWorld/cGrid.cpp
@@ -23,7 +23,7 @@ void cGrid::init(){
 	float _fwidth = m_fLength / m_nGridNum;
 	ST_PC_VERTEX v;
 	for (int i = 0; i < m_nGridNum; i++){
-		v.p = D3DXVECTOR3((0.0f - m_fLength / 2.0f + _fwidth*i), 0, 0.0 + m_fLength / 2.0f);
+		v.p = D3DXVECTOR3(static_cast<float>((0.0f - m_fLength / 2.0f + _fwidth*i)), 0, static_cast<float>(0.0 + m_fLength / 2.0f));
 		if (i % 5 == 0){
 			v.c = D3DCOLOR_XRGB(255, 255, 255);
 		}
@@ -31,12 +31,12 @@ void cGrid::init(){
 			v.c = D3DCOLOR_XRGB(100, 100, 100);
 		}
 		m_vecVertexLineXGrid.push_back(v);
-		v.p = D3DXVECTOR3((0.0f - m_fLength / 2.0f + _fwidth*i), 0, 0.0 - m_fLength / 2.0f);
+		v.p = D3DXVECTOR3(static_cast<float>((0.0f - m_fLength / 2.0f + _fwidth*i)), 0, static_cast<float>(0.0 - m_fLength / 2.0f));
 		m_vecVertexLineXGrid.push_back(v);
 	}
 
 	for (int i = 0; i < m_nGridNum; i++){
-		v.p = D3DXVECTOR3(0.0 + m_fLength / 2.0f, 0, (0.0f - m_fLength / 2.0f + _fwidth*i));
+		v.p = D3DXVECTOR3(static_cast<float>(0.0 + m_fLength / 2.0f), 0, static_cast<float>((0.0f - m_fLength / 2.0f + _fwidth*i)));
 		if (i % 5 == 0){
 			v.c = D3DCOLOR_XRGB(255, 255, 255);
 		}
@@ -44,7 +44,7 @@ void cGrid::init(){
 			v.c = D3DCOLOR_XRGB(100, 100, 100);
 		}
 		m_vecVertexLineZGrid.push_back(v);
-		v.p = D3DXVECTOR3(0.0 - m_fLength / 2.0f, 0, (0.0f - m_fLength / 2.0f + _fwidth*i));
+		v.p = D3DXVECTOR3(static_cast<float>(0.0 - m_fLength / 2.0f), 0, static_cast<float>((0.0f - m_fLength / 2.0f + _fwidth*i)));
 		m_vecVertexLineZGrid.push_back(v);
 	}
 }
